Rename libitina.py to yuri.py to avoid confusion
